requirements.txt
psycopg2
fastapi
uvicorn[standard]
sqlalchemy
gunicorn
python-multipart
python-jose